Reorganize `src` folder
- Move phy-related thing to `src/phy`
- Create separate subfolder for each driver
