+++ src/USBPowerDelivery.h
@@ -15,4 +15,4 @@
 #include "PDController.h"
 #include "PDSink.h"
 #include "PDProtocolAnalyzer.h"
-#include "NucleoSNK1MK1.h"
+#include "./phy/NucleoSNK1MK1/NucleoSNK1MK1.h"
