+++ src/phy/STM32F4/PDPhySTM32F4.h
@@ -8,20 +8,18 @@
 
 #pragma once
 
-#if defined(STM32L4xx)
+#if defined(STM32F401xC)
 
 #include "PDPhy.h"
 
-extern "C" void COMP_IRQHandler();
-extern "C" void ADC1_IRQHandler();
-extern "C" void DMA1_Channel3_IRQHandler();
-extern "C" void DMA2_Channel2_IRQHandler();
+extern "C" void ADC_IRQHandler();
+extern "C" void DMA2_Stream3_IRQHandler();
 
 /**
  * @brief Physical layer for USB PD communication.
- * 
+ *
  */
-struct PDPhySTM32L4 : PDPhy {
+struct PDPhySTM32F4 : PDPhy {
 private:
     enum class RxConfig {
         monitorCC,
@@ -31,23 +29,21 @@ private:
     };
 
     static void initRx();
-    static void initTx();
-    static bool transmitMessage(const PDMessage* msg);
 
     static int decodeChunk(int currentIndex);
     static void processData();
     static void switchCC();
     static void switchAfterDebounce(uint8_t cc);
-    static void messageTransmitted();
     static void configureRx(RxConfig config);
-    static void enablePDCommunication();
-    static void disablePDCommunication();
+    static void onEXTIInterrupt();
+
+    static void initTx();
+    static bool transmitMessage(const PDMessage* msg);
+    static void messageTransmitted();
 
     friend class PDPhy;
-    friend void COMP_IRQHandler();
-    friend void ADC1_IRQHandler();
-    friend void DMA1_Channel3_IRQHandler();
-    friend void DMA2_Channel2_IRQHandler();
+    friend void ADC_IRQHandler();
+    friend void DMA2_Stream3_IRQHandler();
 };
 
 #endif
