+++ src/phy/STM32L4/PDPhySTM32L4.h
@@ -8,18 +8,20 @@
 
 #pragma once
 
-#if defined(STM32F401xC)
+#if defined(STM32L4xx)
 
 #include "PDPhy.h"
 
-extern "C" void ADC_IRQHandler();
-extern "C" void DMA2_Stream3_IRQHandler();
+extern "C" void COMP_IRQHandler();
+extern "C" void ADC1_IRQHandler();
+extern "C" void DMA1_Channel3_IRQHandler();
+extern "C" void DMA2_Channel2_IRQHandler();
 
 /**
  * @brief Physical layer for USB PD communication.
- * 
+ *
  */
-struct PDPhySTM32F4 : PDPhy {
+struct PDPhySTM32L4 : PDPhy {
 private:
     enum class RxConfig {
         monitorCC,
@@ -29,21 +31,23 @@ private:
     };
 
     static void initRx();
+    static void initTx();
+    static bool transmitMessage(const PDMessage* msg);
 
     static int decodeChunk(int currentIndex);
     static void processData();
     static void switchCC();
     static void switchAfterDebounce(uint8_t cc);
-    static void configureRx(RxConfig config);
-    static void onEXTIInterrupt();
-
-    static void initTx();
-    static bool transmitMessage(const PDMessage* msg);
     static void messageTransmitted();
+    static void configureRx(RxConfig config);
+    static void enablePDCommunication();
+    static void disablePDCommunication();
 
     friend class PDPhy;
-    friend void ADC_IRQHandler();
-    friend void DMA2_Stream3_IRQHandler();
+    friend void COMP_IRQHandler();
+    friend void ADC1_IRQHandler();
+    friend void DMA1_Channel3_IRQHandler();
+    friend void DMA2_Channel2_IRQHandler();
 };
 
 #endif
